ojs 2.4.4

+++ classes/journal/JournalDAO.inc.php
@@ -3,8 +3,8 @@
 /**
  * @file classes/journal/JournalDAO.inc.php
  *
- * Copyright (c) 2013 Simon Fraser University Library
- * Copyright (c) 2003-2013 John Willinsky
+ * Copyright (c) 2013-2014 Simon Fraser University Library
+ * Copyright (c) 2003-2014 John Willinsky
  * Distributed under the GNU GPL v2. For full terms see the file docs/COPYING.
  *
  * @class JournalDAO
@@ -145,6 +145,8 @@ class JournalDAO extends DAO {
 	 * @param $journalId int
 	 */
 	function deleteJournalById($journalId) {
+		if (HookRegistry::call('JournalDAO::deleteJournalById', array(&$this, &$journalId))) return;
+
 		$journalSettingsDao =& DAORegistry::getDAO('JournalSettingsDAO');
 		$journalSettingsDao->deleteSettingsByJournal($journalId);
